+++ packages/blocks-testing/test/mock_airtable_interface.test.ts
@@ -113,6 +113,7 @@ describe('MockAirtableInterface', () => {
                     isActive: true,
                 },
             ],
+            workspaceId: 'wsptestworkspaceid',
         };
     });